pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "juntagrico-godparent"
description = "Match godparents with new members in juntagrico to help them get started"
authors = [
    {name = "juntagrico", email = "python@juntagrico.org"},
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "juntagrico>=2.0.0a1",
    "django-multiselectfield==1.0.1"
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Framework :: Django',
    'Framework :: Django :: 5.2',
    'Intended Audience :: Developers',
    'Intended Audience :: Other Audience',
    'License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
    'Topic :: Internet :: WWW/HTTP :: Site Management',
]
dynamic = ["version"]

[project.urls]
Homepage = "https://www.juntagrico.org"
Issues = "https://github.com/juntagrico/juntagrico-godparent/issues"

[tool.setuptools.packages.find]
include = ["juntagrico_godparent*"]

[tool.setuptools.dynamic]
version = {attr = "juntagrico_godparent.__version__"}

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]
